backend/requirements.txt
# Backend
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3

# ML & Data

numpy>=1.23,<2.0
pandas>=1.5,<3.0
scikit-learn>=1.3,<2.0
joblib>=1.3

torch
pydantic>=1.10,<2.0
requests

# NLP / LLM / HuggingFace
torch>=2.0
transformers>=4.36
langchain-google-genai>=1.0.0


# LangChain ecosystem
langchain==0.2.16
langchain-core==0.2.38
langchain-community==0.2.16
langgraph==0.1.19


# Redis Cache
redis>=5.0
django-redis>=5.4

# Database
psycopg2-binary>=2.9

# Utilities
python-dotenv>=1.0
gunicorn>=21.2


# django ui
whitenoise

# CORS
django-cors-headers

# databse_URL
dj-database-url
psycopg2-binary